@@ -1,20 +1,5 @@
 require 'rails_helper'
 
-describe StudentsController do
-
-  describe "POST create" do
-    context "with valid attributes" do
-      it "creates a new student" do
-        expect{
-          post :create, { :student => { :first_name => "Bill", :last_name => "Smith" } }
-        }.to change(Student,:count).by(1)
-      end
-      
-      it "redirects to the new student" do
-        post :create, { :student => { :first_name => "Bill", :last_name => "Smith" } }
-        expect(response).to redirect_to Student.last
-      end
-    end
-  end
+RSpec.describe StudentsController, type: :controller do
 
 end
