@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :school_class do
-    title "Computer Science"
-    room_number 5
+    title { "MyString" }
+    room_number { 1 }
   end
 end
