@@ -1,6 +1,6 @@
-FactoryBot.define do
-  factory :student do
-    first_name "Daenerys"
-    last_name "Targaryen"
-  end
-end
+# FactoryBot.define do
+#   factory :student do
+#     first_name "Daenerys"
+#     last_name "Targaryen"
+#   end
+# end
